Changed exception

+++ src/main/java/com/teamdev/arseniuk/RemoveService.java
@@ -1,6 +1,5 @@
 package com.teamdev.arseniuk;
 
-import java.io.IOException;
 import java.util.Properties;
 import java.util.Set;
 
@@ -19,7 +18,7 @@ public class RemoveService implements Runnable {
         while (!isStopped) {
             try {
                 removeOld();
-            } catch (IOException e) {
+            } catch (FileStorageException e) {
                 e.printStackTrace();
             }
             try {
@@ -30,15 +29,21 @@ public class RemoveService implements Runnable {
         }
     }
 
-    private void removeOld() throws IOException {
+    private void removeOld() throws FileStorageException {
         final FileSystemService fileSystemService = new FileSystemService();
         final Properties expirationDates = systemInformation.getExpirationDates();
         final Set<String> keys = expirationDates.stringPropertyNames();
         for (String key : keys) {
             final String expirationDate = expirationDates.getProperty(key);
+            if(expirationDate == null){
+                continue;
+            }
             final long time = Long.parseLong(expirationDate);
             if (expirationDate != null && time != -1) {
                 Item item = systemInformation.get(key);
+                if (item == null) {
+                    continue;
+                }
                 if ((item.getExpirationTime() + item.getCreationTime()) < System.currentTimeMillis()) {
                     fileSystemService.removeFile(systemInformation.getRootPath() + item.getPath());
                     systemInformation.remove(item);
